Complete Tic Tac Toe game
Allows inviting other users in the same chatroom to play Tic Tac Toe synchronously, with the ability to exit synchronously.
Future enhancements and fixes needed:

There is an error in parameter passing for the restart button. Need to determine the correct parameter passing method.
Display text indicating whose turn it currently is.
Ability to add a scoreboard.
Optimization of the user selection list (remove oneself and automatically hide the button after clicking).